+++ SOC_Photon/src/Coulombs.h
@@ -94,35 +94,35 @@ class Coulombs
 {
 public:
   Coulombs();
-  Coulombs(double *sp_delta_q, float *sp_t_last, const double q_cap_rated, const double t_rated, const double t_rlim,
-    uint8_t *sp_mod_code, const double coul_eff);
+  Coulombs(double *sp_delta_q, float *sp_t_last, const float q_cap_rated, const float t_rated, const float t_rlim,
+    uint8_t *sp_mod_code, const float coul_eff);
   ~Coulombs();
   // operators
   // functions
-  void apply_cap_scale(const double scale);
+  void apply_cap_scale(const float scale);
   void apply_delta_q(const double delta_q);
   void apply_resetting(const boolean resetting){ resetting_ = resetting; };
-  void apply_soc(const double soc, const float temp_c);
+  void apply_soc(const float soc, const float temp_c);
   void apply_delta_q_t(const boolean reset);
   void apply_delta_q_t(const double delta_q, const float temp_c);
   void assign_all_mod(const String mod_str) { chem_.assign_all_mod(mod_str); };
   double calculate_capacity(const float temp_c);
   double coul_eff() { return ( coul_eff_ ); };
   void coul_eff(const double coul_eff) { coul_eff_ = coul_eff; };
-  virtual double count_coulombs(const double dt, const boolean reset, const float temp_c, const double charge_curr,
+  virtual float count_coulombs(const double dt, const boolean reset, const float temp_c, const float charge_curr,
     const boolean sat, const double delta_q_ekf);
   double delta_q() { return(*sp_delta_q_); };
   uint8_t mod_code() { return (*chem_.sp_mod_code); };
   virtual void pretty_print();
   double q(){ return (q_); };
-  double q_cap_rated(){ return (q_cap_rated_); };
-  double q_cap_scaled(){ return (q_cap_rated_scaled_); };
-  double q_capacity(){ return (q_capacity_); };
-  double soc() { return(soc_); };
-  double soc_min() { return(soc_min_); };
+  float q_cap_rated(){ return (q_cap_rated_); };
+  float q_cap_scaled(){ return (q_cap_rated_scaled_); };
+  float q_capacity(){ return (q_capacity_); };
+  float soc() { return(soc_); };
+  float soc_min() { return(soc_min_); };
   boolean sat() { return(sat_); };
-  double t_last() { return(*sp_t_last_); };
-  virtual double vsat(void) = 0;
+  float t_last() { return(*sp_t_last_); };
+  virtual float vsat(void) = 0;
 protected:
   double *sp_delta_q_;// Charge since saturated, C
   float *sp_t_last_;  // Last battery temperature for rate limit memory, deg C
@@ -130,15 +130,15 @@ protected:
   double q_cap_rated_scaled_;// Applied rated capacity at t_rated_, after scaling, C
   double q_capacity_; // Saturation charge at temperature, C
   double q_;          // Present charge available to use, except q_min_, C
-  double soc_;        // Fraction of saturation charge (q_capacity_) available (0-1)
+  float soc_;        // Fraction of saturation charge (q_capacity_) available (0-1)
   boolean sat_;       // Indication that battery is saturated, T=saturated
-  double t_rated_;    // Rated temperature, deg C
-  double t_rlim_;     // Tb rate limit, deg C / s
+  float t_rated_;    // Rated temperature, deg C
+  float t_rlim_;     // Tb rate limit, deg C / s
   boolean resetting_ = false;  // Sticky flag to coordinate user testing of coulomb counters, T=performing an external reset of counter
-  double soc_min_;    // As battery cools, the voltage drops and there appears a minimum soc it can deliver
-  double q_min_;      // Floor on charge available to use, C
+  float soc_min_;    // As battery cools, the voltage drops and there appears a minimum soc it can deliver
+  float q_min_;      // Floor on charge available to use, C
   Chemistry chem_; // Storage of chemistry information
-  double coul_eff_;   // Coulombic efficiency - the fraction of charging input that gets turned into usable Coulombs
+  float coul_eff_;   // Coulombic efficiency - the fraction of charging input that gets turned into usable Coulombs
 };
 
 #endif
